break down context file
